+++ src/Console/ModelsCommand.php
@@ -106,6 +106,8 @@ class ModelsCommand extends Command
     protected $keep_text;
     protected $phpstorm_noinspections;
     protected $write_model_external_builder_methods;
+    protected $phpdoc_separate_tags = false;
+
     /**
      * @var bool[string]
      */
@@ -158,6 +160,7 @@ class ModelsCommand extends Command
         $this->write_model_external_builder_methods = $this->laravel['config']->get('ide-helper.write_model_external_builder_methods', true);
         $this->write_model_relation_count_properties =
             $this->laravel['config']->get('ide-helper.write_model_relation_count_properties', true);
+        $this->phpdoc_separate_tags = $this->laravel['config']->get('ide-helper.phpdoc_separate_tags', false);
 
         $this->write = $this->write_mixin ? true : $this->write;
         //If filename is default and Write is not specified, ask what to do
@@ -1009,7 +1012,7 @@ class ModelsCommand extends Command
             );
         }
 
-        $serializer = new DocBlockSerializer();
+        $serializer = new DocBlockSerializer(0, ' ', true, null, $this->phpdoc_separate_tags);
         $docComment = $serializer->getDocComment($phpdoc);
 
         if ($this->write_mixin) {