+++ src/Eloquent.php
@@ -70,7 +70,8 @@ class Eloquent
             return;
         }
 
-        $serializer = new DocBlockSerializer();
+        $separateTags = $command->getLaravel()['config']->get('ide-helper.phpdoc_separate_tags', false);
+        $serializer = new DocBlockSerializer(0, ' ', true, null, $separateTags);
         $serializer->getDocComment($phpdoc);
         $docComment = $serializer->getDocComment($phpdoc);