+++ CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Add support for nikic/php-parser 5 (next to 4) [#1502 / mfn](https://github.com/barryvdh/laravel-ide-helper/pull/1502)
 - Add support for `immutable_date:*` and `immutable_datetime:*` casts. [#1380 / thekonz](https://github.com/barryvdh/laravel-ide-helper/pull/1380)
+- Added separation of phpdoc tags into groups [#1377 / Kerigard](https://github.com/barryvdh/laravel-ide-helper/pull/1377)
 
 2023-02-04, 2.13.0
 ------------------